requirements.txt
bcrypt==3.1.0
django-cors-headers==1.1.0
inflection==0.3.1
jwt==0.3.2
pydash==3.4.3
PyJWT==1.4.2
python-dateutil==2.5.3
requests==2.11.0
